backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.5",
    "fastapi>=0.135.1",
    "httpx>=0.28.0",
    "PyJWT>=2.10.0",
    "litellm>=1.82.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "supabase>=2.28.0",
    "uvicorn[standard]>=0.41.0",
]
